+++ Buzzer_Windows_App/Buzzer_Windows_App/UmActually/UmActually.xaml.cs
@@ -27,7 +27,7 @@ namespace Buzzer_Windows_App
         PlayerContainer[] players;
 
         bool HasBuzzed = false;
-        uint BuzzedPlayer = 0;
+        UInt64 BuzzedPlayer = 0;
 
         public UmActually(ControllerSettingsContainer controllers, MessageDecoder decoder, SettingsContainer settings)
         {
@@ -46,7 +46,7 @@ namespace Buzzer_Windows_App
 
         }
 
-        private void M_decoder_OnButtonPressed(object sender, uint e)
+        private void M_decoder_OnButtonPressed(object sender, UInt64 e)
         {
             if (HasBuzzed)
                 return;
@@ -140,7 +140,7 @@ namespace Buzzer_Windows_App
                 m_score.MouseRightButtonDown -= DecreaseScore;
             }
 
-            public uint GetID()
+            public UInt64 GetID()
             {
                 return m_player.ID;
             }
